% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{calc_CWD_mm}
\alias{calc_CWD_mm}
\title{Climatic water deficit}
\usage{
calc_CWD_mm(pet_cm, et_cm)
}
\arguments{
\item{pet_cm}{A numeric vector. Potential evapotranspiration \verb{[cm]}.}

\item{et_cm}{A numeric vector. Actual evapotranspiration \verb{[cm]}.}
}
\value{
Climatic water deficit \verb{[mm]}.
}
\description{
Climatic water deficit
}
