% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_004_Recruitment.R
\name{calc_RecruitmentIndex_v3}
\alias{calc_RecruitmentIndex_v3}
\title{Index estimating recruitment potential for perennial plants}
\usage{
calc_RecruitmentIndex_v3(
  sim_data,
  soils,
  out = c("ts_years", "raw"),
  hemisphere_NS = c("N", "S"),
  recruitment_depth_range_cm = c(5, 30),
  Temp_limit_C = 5,
  Wet_SWP_limit_MPa = -1.5,
  Dry_SWP_limit_MPa = -3,
  init_WDD = 15,
  init_days = 3,
  init_depth_range_cm = c(0, 5),
  stop_DDD = 0,
  stop_days_DDD = 0,
  stop_depth_range_cm = c(0, 30),
  stop_TDD = 0,
  stop_days_TDD = 0,
  include_year = FALSE,
  tol = sqrt(.Machine[["double.eps"]]),
  ...
)
}
\arguments{
\item{sim_data}{A named list or environment with the following elements
\itemize{
\item \code{"swp_daily"} with two elements
\itemize{
\item \code{"values"}, a list with the named element \code{"swp"},
a two-dimensional object (days by soil layers)
of daily soil water potential \verb{[-bar]}
\item \code{"time"}, a two-dimensional object with days in rows and columns
\code{"Year"} and \code{"Day"} (day of year)
}
\item \code{"temp_daily"}, a list with the element \code{"values"} that contains the
named element \code{"tmean"}, a vector of daily mean air temperatures \verb{[C]}
\item \code{"swe_daily"}, a list with the named element\code{"values"} that contains the
named element \code{"swe"}, a vector of daily water equivalents of
the snow pack \verb{[cm]}
}}

\item{soils}{A named list with soil parameters \code{"depth_cm"},
\code{"sand_frac"}, \code{"clay_frac"}, and \code{"gravel_content"}
as numeric vectors with values for each soil layer.}

\item{out}{A character string. Signaling whether the functions returns
a time series of yearly values or an aggregate (e.g., mean) across years.
One of \var{\dQuote{ts_years}} or \var{\dQuote{across_years}}.}

\item{hemisphere_NS}{A character. Hemisphere identifies day of mid-summer
(July 15 on the northern hemisphere) that separates spring and fall
recruitment periods.}

\item{recruitment_depth_range_cm}{A numeric vector of length two. Soil depth
interval over which wet degree-days are accumulated.}

\item{Temp_limit_C}{A numeric value. The base temperature used to accumulate
degree-days.}

\item{Wet_SWP_limit_MPa}{A numeric value. Critical soil water potential that
identifies moist soil conditions.}

\item{Dry_SWP_limit_MPa}{A numeric value. Critical soil water potential that
identifies dry soil conditions.}

\item{init_WDD}{A numeric value.}

\item{init_days}{A numeric value.}

\item{init_depth_range_cm}{A numeric vector of length two.}

\item{stop_DDD}{A numeric value.}

\item{stop_days_DDD}{A numeric value.}

\item{stop_depth_range_cm}{A numeric vector of length two.}

\item{stop_TDD}{A numeric value.}

\item{stop_days_TDD}{A numeric value.}

\item{include_year}{A logical value.
A column \code{"Year"} is added to the output.}

\item{tol}{A numeric value.}

\item{...}{Additional arguments}
}
\description{
Recruitment potential is estimated as the sum of wet degree-days \code{WDD}
during the most favorable continuous periods with warm conditions and
wet near-surface soils, see \code{\link[=calc_MDD_daily]{calc_MDD_daily()}}.
The function calculates the onset timing, duration and accumulated \code{WDD}
for the most favorable period in the spring and fall,
see Chenoweth et al. (2023) for additional details.
}
\details{
Specifically, recruitment potential considers \code{WDD} over soil depths of
\code{recruitment_depth_range_cm} during periods that are defined
by starting and stopping conditions.

The start of a suitable period is identified by a wet period (positive \code{WDD})
that lasts at least \code{init_days} and accumulated \code{WDD} over soil depths of
\code{init_depth_range_cm} reaches at least \code{init_WDD}.

The end of a suitable period is identified either
(i) by a dry period (positive dry degree-days \code{DDD})
that lasts at least \code{stop_days_DDD} and accumulated \code{DDD} over soil depths of
\code{stop_depth_range_cm} reaches at least \code{stop_DDD}, or
(ii) by a period with accumulated total degree-days \code{TDD} of
less than \code{stop_TDD} that lasts at least \code{stop_days_TDD}.
}
\section{Details}{

Argument \code{soils} uses only element \code{"depth_cm"}.
}

\examples{
# Prepare data (here using rSOILWAT2)
swin <- rSOILWAT2::sw_exampleData
soils <- list(depth_cm = rSOILWAT2::swSoils_Layers(swin)[, "depth_cm"])
nSoilLayers <- length(soils[["depth_cm"]])

sim <- rSOILWAT2::sw_exec(swin)
sim_data <- list(
  swp_daily = list(
    time = sim@SWPMATRIC@Day[, c("Year", "Day")],
    values = list(
      swp = sim@SWPMATRIC@Day[, paste0("Lyr_", seq_len(nSoilLayers))]
    )
  ),
  temp_daily = list(
    values = list(tmean = sim@TEMP@Day[, "avg_C"])
  ),
  swe_daily = list(
    values = list(swe = sim@SNOWPACK@Day[, "snowpackWaterEquivalent_cm"])
  )
)

# Recruitment potential
ri <- calc_RecruitmentIndex_v3(
  sim_data = sim_data,
  soils = soils,
  recruitment_depth_range_cm = c(10, 20),
  Temp_limit_C = 5,
  Wet_SWP_limit_MPa = -1.5,
  Dry_SWP_limit_MPa = -3,
  init_WDD = 15,
  init_days = 3,
  init_depth_range_cm = c(0, 10),
  stop_DDD = 15,
  stop_days_DDD = 3,
  stop_depth_range_cm = c(0, 20),
  stop_TDD = 0,
  stop_days_TDD = 3,
  include_year = TRUE
)

}
\references{
Chenoweth et al. (2023)
Ecologically relevant moisture and temperature metrics for assessing
dryland ecosystem dynamics.
Ecohydrology, 16(3), e2509. \url{https://doi.org/10.1002/eco.2509}
}
