% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_009_DailyClim.R
\name{calc_SWP_MPa}
\alias{calc_SWP_MPa}
\title{Determine soil water potential \var{SWP}}
\usage{
calc_SWP_MPa(
  sim_swc_daily,
  time,
  soils,
  used_depth_range_cm = NULL,
  method = c("across_profile", "by_layer"),
  out = c("ts_years", "across_years"),
  fun_aggs_across_yrs = mean
)
}
\arguments{
\item{sim_swc_daily}{A numeric two-dimensional object with
daily \var{"swc"} for each soil layer in units of centimeters.}

\item{time}{A numeric vector. Time for each row of \code{sim_swc_daily}.}

\item{soils}{A named list with soil parameters \code{"depth_cm"},
\code{"sand_frac"}, \code{"clay_frac"}, and \code{"gravel_content"}
as numeric vectors with values for each soil layer.}

\item{used_depth_range_cm}{A numeric vector of length two (or \code{NULL}).
The upper and lower depth limit of the zone (depth range) to consider.}

\item{method}{A character string.}

\item{out}{A character string.}

\item{fun_aggs_across_yrs}{A function that calculates across-year
summaries that returns a named or unnamed vector.}
}
\value{
A list with elements "time" and "values" where values represent
soil water potential in units \code{MPa}:
\itemize{
\item if \code{method} is \var{\dQuote{across_profile}},
then summed across \code{used_depth_range_cm},
\item if \code{method} is \var{\dQuote{by_layer}},
then columns contain values for each soil layer within
\code{used_depth_range_cm};
\item if \code{out} is \verb{"ts_years},
then rows contain values for each time step,
\item if \code{out} is \code{"across_years"},
then rows contain averaged values across unique time steps.
}
}
\description{
Determine soil water potential \var{SWP}
}
\section{Details}{

After soil moisture has been aggregated across soil layers
if requested (\code{method} is \code{"across_profile"}) and
aggregated across years if requested (\code{out} is \code{"across_years"}), then
the water release curve employed during the simulation run is used to
translate volumetric water content \code{VWC} into \code{SWP}, i.e.,
\deqn{SWP[i,matric] = f(VWC, SWRCp[i])}
where \code{SWRCp} are the parameters describing the water release curve.
Note, the translation is only valid for the matric soil, i.e.,
the component without coarse fragments.
}

