% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{calc_univariate_from_sw2}
\alias{calc_univariate_from_sw2}
\title{\var{SEUG}-era extractor of \pkg{rSOILWAT2} output data}
\usage{
calc_univariate_from_sw2(
  path,
  name_sw2_run,
  id_scen_used,
  list_years_scen_used,
  group_by_month,
  first_month_of_year,
  req_ts = TRUE,
  sw2_tp,
  sw2_out,
  sw2_var,
  varnames_are_fixed = TRUE,
  fun_across_vars,
  fun_across_time,
  ...
)
}
\value{
A list of matrices.
  Each matrix contains the output for one \code{id_scen_used};
  rows represent the aggregation groups \code{group_by_month}, and
  columns represent time, e.g., calendar years.
}
\description{
\var{SEUG}-era extractor of \pkg{rSOILWAT2} output data
}
