% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{check_soillayer_availability}
\alias{check_soillayer_availability}
\title{Check whether soil layers are available in a specified zone}
\usage{
check_soillayer_availability(
  soil_depths_cm,
  used_depth_range_cm = NULL,
  strict = TRUE,
  type = c("warn", "error")
)
}
\arguments{
\item{soil_depths_cm}{A numeric vector.
The lower depth limits of soil layers in \var{[cm]}.}

\item{used_depth_range_cm}{A numeric vector of length two (or \code{NULL}).
The upper and lower depth limit of the zone (depth range) to consider.}

\item{strict}{A logical vector either of length one or equal to the
length of \code{used_depth_range_cm}. \code{TRUE} indicates that
the corresponding value (or all) of \code{used_depth_range_cm} is required
to be either zero or one of the values of \code{soil_depths_cm}.}

\item{type}{A character string. If a value is not present, then either
a warning is issued or an error.}
}
\value{
(Invisibly) a logical value indicating whether at least one
required value is missing (\code{FALSE}) or all are present (\code{TRUE}).
}
\description{
Check whether soil layers are available in a specified zone
}
\examples{
check_soillayer_availability(
  c(5, 10, 30, 50), c(0, 40),
  strict = TRUE,
  type = "warn"
)
check_soillayer_availability(
  c(5, 10, 30, 50), c(0, 40),
  strict = c(TRUE, FALSE),
  type = "warn"
)
check_soillayer_availability(
  c(5, 10, 30, 50), c(20, 30, 40),
  strict = TRUE,
  type = "warn"
)

}
