% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{convert_with_swrc}
\alias{convert_with_swrc}
\title{Convert between \code{VWC} and \code{SWP} using a soil water retention curve}
\usage{
convert_with_swrc(
  x,
  direction = c("vwc_to_swp", "swp_to_vwc"),
  use_swrc_v6 = getNamespaceVersion("rSOILWAT2") >= numeric_version("6.0.0"),
  fcoarse = NULL,
  sand = NULL,
  clay = NULL,
  swrcp = NULL,
  swrc_name = NULL
)
}
\arguments{
\item{x}{A numeric object with values of either
volumetric water content or of soil water potential.}

\item{direction}{A text string.}

\item{use_swrc_v6}{A logical value whether the new \code{SWRC} functionality
with \code{rSOILWAT2} \verb{>= v6.0.0} can be used, see
\code{\link[=load_swrcp_and_usage]{load_swrcp_and_usage()}}.}

\item{fcoarse}{A numeric vector. Set to zero values if \code{vwc} represent
the matric component.}

\item{sand}{A numeric vector.}

\item{clay}{A numeric vector.}

\item{swrcp}{A numeric matrix with parameters of the \code{SWRC},
see \code{\link[rSOILWAT2:SWRCs]{rSOILWAT2::SWRCs}}.}

\item{swrc_name}{A text string; the name of the \code{SWRC},
see \code{\link[rSOILWAT2:swrc_names]{rSOILWAT2::swrc_names()}}.}
}
\description{
Convert between \code{VWC} and \code{SWP} using a soil water retention curve
}
\section{Details}{

Available functionality depends on the currently installed version
of \code{rSOILWAT2} and of the version of \code{rSOILWAT2} that was used
to create the simulation input object \code{sw_in}, see
\code{\link[=load_swrcp_and_usage]{load_swrcp_and_usage()}}.


If \code{use_swrc_v6} is \code{TRUE}, then \code{fcoarse}, \code{swrcp}, and \code{swrc_name}
are required arguments.


If \code{use_swrc_v6} is \code{FALSE}, then \code{sand} and \code{clay} are required
arguments.
}

\seealso{
\code{\link[rSOILWAT2:swrc_conversion]{rSOILWAT2::swrc_conversion()}}
}
