% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{determine_sw2_sim_time}
\alias{determine_sw2_sim_time}
\title{Prepare specified vs. requested time step \code{data.frame}}
\usage{
determine_sw2_sim_time(
  xt,
  req_years = NULL,
  sw2_tp = c("Day", "Month", "Year")
)
}
\arguments{
\item{xt}{A two-dimensional object with columns
\code{Year},
\code{Month} (if \code{sw2_tp} equals "Month"), and
\code{Day} (if \code{sw2_tp} equals "Day"). Each row represents one time step
according to \code{sw2_tp}.}

\item{req_years}{An integer vector of Calendar years. If missing or \code{NULL},
then time steps contained in \code{xt} are used.
If not missing and not \code{NULL} and different than years contained in \code{xt},
then time steps combined from requested and simulated are used.}

\item{sw2_tp}{A character string. The daily, monthly, or yearly time step
describing content of \code{xt} and determining the same output time step.}
}
\value{
A \code{data.frame} with four columns \code{Year}, \code{Month}, \code{Day}, \code{mode}.
Rows represent daily, monthly, or yearly time steps that combine
requested (if \code{req_years} was provided) and simulated (\code{xt}) time steps.
\code{mode} is \code{"nosim"} if a time step was not simulated (but requested),
\code{"sim_keep"} if a time step was simulated
(and requested or \code{req_years} is missing), or
\code{"sim_discard"} if a time step was simulated (but not requested).
}
\description{
Prepare specified vs. requested time step \code{data.frame}
}
\section{Notes}{

The column \code{Day} contains day of year (and not day of month)!
}

\examples{
 xty <- data.frame(Year = 1991:2020)
 determine_sw2_sim_time(xty, sw2_tp = "Year")
 determine_sw2_sim_time(xty, req_years = 1981:2010, sw2_tp = "Year")
 determine_sw2_sim_time(xty, req_years = 2011:2030, sw2_tp = "Year")

}
