% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{EcologicalDroughtMetrics2023_annualClim}
\alias{EcologicalDroughtMetrics2023_annualClim}
\alias{metric_EcologicalDroughtMetrics2023_annualClim}
\title{Climatologies of annual time series of ecological drought metrics}
\usage{
metric_EcologicalDroughtMetrics2023_annualClim(
  path,
  name_sw2_run,
  id_scen_used,
  list_years_scen_used,
  out = "across_years",
  zipped_runs = FALSE,
  soils,
  ...
)
}
\value{
A return object where \code{group} contains the following climatologies of
annual variables summarized across years by means \code{mean},
standard deviations \code{sd}, coefficients of variation \code{cv}, or
frequency of occurrence \code{frq}:
\itemize{
\item Aridity index \verb{[mm / mm]}: \code{"AI_(mean)|(cv)"}
(where \code{AI = PPT / PET})
\item Potential evapotranspiration amount \verb{[mm]}: \code{"PET_(mean)|(cv)"}
\item Mean daily air temperature \verb{[C]}: \code{"Tmean_(mean)|(sd)"}
\item Minimum daily air temperature \verb{[C]}: \code{"Tmin_(mean)|(sd)"}
\item Maximum daily air temperature \verb{[C]}: \code{"Tmax_(mean)|(sd)"}
\item Precipitation amount \verb{[mm]}: \code{"PPT_(mean)|(cv)"}
\item Seasonal timing of precipitation \verb{[-]}: \code{"PPTsst_(mean)|(sd)"}
(where \verb{PPTsst = cor(monthly PPT, monthly Tmean)})
\item Total growing degree days \verb{[C x day]}: \code{"TDD_(mean)|(cv)"}
\item Seasonal variability of total growing degree days \verb{[C x day / C x day]}: \code{"TDDssv_(mean)|(cv)"}
(where \verb{TDDssv = mean(monthly TDD) / sd(monthly TDD)})
\item Warm season length \verb{[day]}: \code{"GrowingSeasonDuration_(mean)|(cv)"}
(where \code{GrowingSeasonDuration} is the longest spell of days with a positive \code{TDD})
\item Evapotranspiration amount \verb{[mm]}: \code{"ET_(mean)|(cv)"}
\item Climatic water deficit amount \verb{[mm]}: \code{"CWD_(mean)|(cv)"}
(where \code{CWD = PPT - ET})
\item 10-day extreme climatic water deficit \verb{[mm]}: \code{"CWDextreme010d_(mean)|(cv)"}
(where \code{CWDextreme10d} is the largest daily \code{CWD} averaged over 10-day periods)
\item Seasonal variability of climatic water deficit \verb{[mm / mm]}: \code{"CWDssv_(mean)|(cv)"}
(where \verb{CWDssv = mean(monthly CWD) / sd(monthly CWD)})
\item Seasonal timing of climatic water deficit \verb{[-]}: \code{"CWDsst_(mean)|(sd)"}
(where \verb{CWDsst = cor(monthly CWD, monthly Tmean)})
\item Dry-degree days (0-100 cm, <-3 MPa) \verb{[C x day]}: \code{"DDD_(mean)|(cv)"}
\item Dry-degree days of longest spell \verb{[C x day]}: \code{"DDDmaxSpell_(mean)|(cv)"}
\item Wet-degree days (0-100 cm, any >-1.5 MPa) \verb{[C x day]}: \code{"WDD_(mean)|(cv)"}
\item Seasonal timing of wet-degree days \verb{[-]}: \code{"WDDsst_(mean)|(cv)"}
(where \verb{WDDsst = cor(monthly WDD, monthly Tmean)})
\item Mean spell length of dry soils (0-100 cm, <-1.5 MPa) \verb{[day]}: \code{"DSI_(mean)|(cv)"}
\item Number of dry soils intervals \verb{[#]}: \code{"nDSI_(mean)|(cv)"}
\item First fall frost (<-5 C) \verb{[day of year]}: \code{"FrostFallFirst_(mean)|(sd)"}
Also, frequency of years with fall frost events \verb{[# / #]}: \code{"FrostFall_frq"}
\item Last spring frost (<-5 C) \verb{[day of year]}: \code{"FrostSpringLast_(mean)|(sd)"}
Also, frequency of years with spring frost events \verb{[# / #]}: \code{"FrostSpring_frq"}
\item Fall recruitment onset \verb{[day of year]}: \code{"RecruitmentFallOnset_(mean)|(sd)"}
Also, frequency of years with fall recruitment \verb{[# / #]}: \code{"RecruitmentFall_frq"}
\item Fall recruitment duration \verb{[day]}: \code{"RecruitmentFallDuration_(mean)|(cv)"}
\item Fall recruitment wet-degree days \verb{[C x day]}: \code{"RecruitmentFallWDD_(mean)|(cv)"}
\item Spring recruitment onset \verb{[day of year]}: \code{"RecruitmentSpringOnset_(mean)|(sd)"}
Also, frequency of years with spring recruitment \verb{[# / #]}: \code{"RecruitmentSpring_frq"}
\item Spring recruitment duration \verb{[day]}: \code{"RecruitmentSpringDuration_(mean)|(cv)"}
\item Spring recruitment wet-degree days \verb{[C x day]}: \code{"RecruitmentSpringWDD_(mean)|(cv)"}
\item Available soil moisture (0-100 cm, >-3.9 MPa) \verb{[mm]}: \code{"SWA_(mean)|(cv)"}
\item Seasonal variability of available soil moisture \verb{[mm / mm]}: \code{"SWAssv_(mean)|(cv)"}
(where \verb{SWAssv = mean(monthly SWA) / sd(monthly CWD)})
\item Seasonal timing of available soil moisture \verb{[-]}: \code{"SWAsst_(mean)|(sd)"}
(where \verb{SWAsst = cor(monthly SWA, monthly Tmean)})
}
}
\description{
Long-term means, standard deviations, or coefficients of variation of
annual values of metrics used by Chenoweth et al.
(see \code{metric_EcologicalDroughtMetrics2023_annual()}).
}
\section{Notes}{

\itemize{
\item Argument \code{fun_aggs_across_yrs} is ignored.
\item Values are \code{NA} if any year is requested but was not simulated.
}
}

\references{
Chenoweth et al. (2023)
Ecologically relevant moisture and temperature metrics for assessing
dryland ecosystem dynamics.
Ecohydrology, 16(3), e2509. \url{https://doi.org/10.1002/eco.2509}
}
\seealso{
\code{\link[=EcologicalDroughtMetrics2023_annual]{EcologicalDroughtMetrics2023_annual()}} and
\code{\link[=metric_EcologicalDroughtMetrics2023_annualClim]{metric_EcologicalDroughtMetrics2023_annualClim()}}
}
