% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractions_Main.R
\name{format_metric_1sim}
\alias{format_metric_1sim}
\title{Transform result into wide format suitable for spreadsheet concatenation}
\usage{
format_metric_1sim(x, id)
}
\description{
Transform result into wide format suitable for spreadsheet concatenation
}
