% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_AcrossYears.R
\name{fun_across_years}
\alias{fun_across_years}
\alias{mean_cv}
\alias{mean_cv_trend}
\alias{mean_cv_mmk}
\alias{mean_sd_cv_mmk}
\title{Useful functions for aggregate statistics of values across years}
\usage{
mean_cv(x, na.rm = TRUE, ...)

mean_cv_trend(x, na.rm = TRUE, ...)

mean_cv_mmk(x, na.rm = TRUE, ...)

mean_sd_cv_mmk(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical value}

\item{...}{Additional parameters passed to the function(s).}
}
\description{
Useful functions for aggregate statistics of values across years
}
