% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{get_values_from_sw2}
\alias{get_values_from_sw2}
\title{Extract values, determine time series, and assign time steps to groups
based on an \pkg{rSOILWAT2} output as generated by \pkg{rSFSW2}}
\usage{
get_values_from_sw2(
  id_scen,
  path,
  name_sw2_run,
  group_by_month,
  first_month_of_year,
  sw2_tp,
  sw2_out,
  sw2_var,
  varnames_are_fixed = TRUE
)
}
\arguments{
\item{id_scen}{An integer value. The climate scenario ID as generated by
\pkg{rSFSW2}.}

\item{path}{A character string. The directory path where \pkg{rSFSW2}
stored the \pkg{rSOILWAT2} output, by default \dQuote{3_Runs}.}

\item{name_sw2_run}{A character string. The \pkg{rSFSW2} name identifying
a run, i.e., the full experimental label x site name. The folder name
within \code{path}.}

\item{group_by_month}{An integer vector of length 12. The group IDs of each
month. If no grouping is requested, then provide,
e.g., \code{rep(0, 12)}.}

\item{first_month_of_year}{An integer value. The number of the month [1-12]
when the "year" should start, e.g., \code{10} for "water-year".
Note: incomplete years, i.e., the first and last year will be discarded.}

\item{sw2_tp}{A character string. One of the \pkg{rSOILWAT2} output time
slots. Note: currently, only \dQuote{Month} and \dQuote{Day} are
implemented.}

\item{sw2_out}{A character vector. One (or multiple) of the \pkg{rSOILWAT2}
output groups. These will be recycled if \code{sw_var} is longer.}

\item{sw2_var}{A character vector. One (or multiple) of the column names
of the output group(s) corresponding to the \code{sw2_out}.}

\item{varnames_are_fixed}{A logical value. If \code{TRUE}, then
\code{sw2_var} is taken literally. Otherwise, \code{sw2_var}
are interpreted as regular expressions
(and may thus return multiple columns).}
}
\value{
A list with three elements: \itemize{
  \item \var{vals} A list with one element for each \code{sw2_var}. Each
        element is a vector of values for that \code{sw2_var},
        if \code{varnames_are_fixed} is \code{TRUE}; otherwise, a vector
        or matrix where columns represent all the \pkg{SOILWAT2} output
        corresponding to that \code{sw2_var} interpreted as
        regular expression.
        The length of vectors (or the number of rows) corresponds to the
        number time steps requested by \code{sw2_tp}.
  \item \var{year} A vector of calendar years for each time step.
  \item \var{groups_by_time} A vector of group ID, based on
        \code{group_by_month}, for each time step.
}
}
\description{
Extract values, determine time series, and assign time steps to groups
based on an \pkg{rSOILWAT2} output as generated by \pkg{rSFSW2}
}
