% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2metrics-package.R
\name{inputcollectors}
\alias{inputcollectors}
\alias{collect_input_soillayers_clay}
\alias{collect_input_soillayers_count}
\alias{collect_input_soillayers_depth}
\alias{collect_input_soillayers_gravel}
\alias{collect_input_soillayers_sand}
\alias{input}
\alias{inputs}
\alias{collect_input}
\alias{collectors}
\title{End-user functions that collect a specific \code{input}}
\arguments{
\item{path}{A character string. The path to the simulation project folder
that contains the individual folder of each simulated site.}

\item{name_sw2_run}{A character string. The name of the folder of the
simulated site for which metrics are to be calculated.
\pkg{rSOILWAT2} input and output is organized following conventions of
\pkg{rSFSW2}, i.e., inputs for each scenario are stored in a list object
named \var{\dQuote{swRunScenariosData}} which is stored
on disk as a file \var{\dQuote{sw_input.RData}};
and output data is stored for each scenario separately in an object
named \var{\dQuote{runDataSC}}
in a file \var{\dQuote{sw_output_scX.RData}} where
\code{X} is the number of the scenario.}

\item{...}{Additional arguments}
}
\description{
These functions have a name that starts with \code{collect_input_}.
They have at least the following arguments:
\code{path}, \code{name_sw2_run}, and \code{...}.
}
\section{List of currently available input collectors:}{
 \itemize{
   \item \var{collect_input_soillayers_clay}
  \item \var{collect_input_soillayers_count}
  \item \var{collect_input_soillayers_depth}
  \item \var{collect_input_soillayers_gravel}
  \item \var{collect_input_soillayers_sand} 
}}
