% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Selfdescribing.R
\name{list_all_input_collectors}
\alias{list_all_input_collectors}
\title{List all input collecting functions}
\usage{
list_all_input_collectors()
}
\description{
List all input collecting functions
}
