% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{load_swrcp_and_usage}
\alias{load_swrcp_and_usage}
\title{Load soil water retention curve information from a \code{rSOILWAT2} input object}
\usage{
load_swrcp_and_usage(
  x,
  vars = c(depth_cm = "depth_cm", `bulkDensity_g/cm^3` = NA_character_, gravel_content =
    "gravel_content", sand_frac = "sand_frac", clay_frac = "clay_frac")
)
}
\arguments{
\item{x}{A \code{rSOILWAT2} input object of class \code{"swInputData"} or
a data frame (see notes).}

\item{vars}{A named character vector. Column names of \code{x} that are required
to estimate parameters of \code{SWRC} (only used if \code{x} is a data frame).}
}
\value{
A named list including
\itemize{
\item \code{"use_swrc_v6"}, a logical vector that is \code{TRUE}
if the installed version of the \code{rSOILWAT2} package is version \verb{>= 6.0.0}
and if the the simulation input object \code{x} was created by \code{rSOILWAT2}
version \verb{>= 6.0.0} (or if it can be updated to such a version);
\item \code{"swrc_name"}, the name of the soil water retention curve selected in
\code{x} (only used if \code{"use_swrc_v6"}), see \code{\link[rSOILWAT2:swrc_names]{rSOILWAT2::swrc_names()}};
\item \code{"swrcp"}, a matrix with parameters of the selected soil water retention
curve (only used if \code{"use_swrc_v6"}), see \code{\link[rSOILWAT2:SWRCs]{rSOILWAT2::SWRCs}}.
}
}
\description{
Available functionality depends on the currently installed version
of \code{rSOILWAT2} and of the version of \code{rSOILWAT2} that was used
to create the simulation input object \code{sw_in}.
}
\section{Notes}{

If \code{x} is a data frame, then rows represent soil layers and
columns represent soil properties.
Column names are provided via argument \code{vars}.
\code{SWRC} is assumed to be \code{"Campbell1974"} and
the utilized \code{PTF} is \code{"Cosby1984AndOthers"}.
}

\examples{
res1 <- load_swrcp_and_usage(rSOILWAT2::sw_exampleData)
res2 <- load_swrcp_and_usage(
  rSOILWAT2::swSoils_Layers(rSOILWAT2::sw_exampleData)
)
all.equal(res1, res2)

}
\seealso{
\code{\link[=convert_with_swrc]{convert_with_swrc()}}
}
