% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{metric_DR}
\alias{metric_DR}
\title{Diffuse Recharge (DR) = Deep Drainage [mm]}
\usage{
metric_DR(
  path,
  name_sw2_run,
  id_scen_used,
  list_years_scen_used,
  out = "ts_years",
  ...
)
}
\description{
Diffuse Recharge (DR) = Deep Drainage [mm]
}
