% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{metric_ET}
\alias{metric_ET}
\title{Evapotranspiration (ET) [mm]}
\usage{
metric_ET(
  path,
  name_sw2_run,
  id_scen_used,
  list_years_scen_used,
  out = "ts_years",
  ...
)
}
\description{
Evapotranspiration (ET) [mm]
}
