% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{metric_Radiation}
\alias{metric_Radiation}
\title{Radiation (H) [MJ/m2]}
\usage{
metric_Radiation(
  path,
  name_sw2_run,
  id_scen_used,
  list_years_scen_used,
  out = "ts_years",
  ...
)
}
\description{
Radiation (H) [MJ/m2]
}
