% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2metrics-package.R
\name{metrics}
\alias{metrics}
\alias{metric_AI}
\alias{metric_CWD}
\alias{metric_Climate_annual}
\alias{metric_Climate_monthly}
\alias{metric_Climate_quarterly}
\alias{metric_CorTP_Annual}
\alias{metric_CorTempPPT}
\alias{metric_DDDat5C0to020cm30bar_quarterly}
\alias{metric_DDDat5C0to030cm30bar}
\alias{metric_DDDat5C0to100cm30bar}
\alias{metric_DDDat5C0to100cm30bar_quarterly}
\alias{metric_DR_JJA}
\alias{metric_DR_annual}
\alias{metric_DR_daily}
\alias{metric_DR_monthly}
\alias{metric_DSIat0to100cm15bar}
\alias{metric_DSIat0to100cm30bar}
\alias{metric_DrySoilDays_Seasonal_top50cm}
\alias{metric_DrySoilDays_Seasonal_wholeprofile}
\alias{metric_ET_annual}
\alias{metric_ET_monthly}
\alias{metric_Evaporation_Seasonal}
\alias{metric_ExtremeShortTermDryStress_Seasonal_top50cm}
\alias{metric_ExtremeShortTermDryStress_Seasonal_wholeprofile}
\alias{metric_FrostDaysAtNeg5C}
\alias{metric_FrostDays_Seasonal}
\alias{metric_NonDrySWA_Seasonal_top50cm}
\alias{metric_NonDrySWA_Seasonal_wholeprofile}
\alias{metric_PET_Seasonal}
\alias{metric_PPT_JJA}
\alias{metric_PPT_Seasonal}
\alias{metric_PPT_daily}
\alias{metric_PPT_dailyClim}
\alias{metric_PPT_monthlyClim}
\alias{metric_Radiation_annual}
\alias{metric_Radiation_monthly}
\alias{metric_RecruitmentIndex_v4}
\alias{metric_RecruitmentIndex_v5}
\alias{metric_SMTRs}
\alias{metric_SWA_Seasonal_top50cm}
\alias{metric_SWA_Seasonal_wholeprofile}
\alias{metric_SWAat0to020cm39bar_JJA}
\alias{metric_SWAat0to020cm39bar_daily}
\alias{metric_SWAat0to020cm39bar_dailyClim}
\alias{metric_SWAat0to020to100cm39bar_quarterly}
\alias{metric_SWAat0to100cm30bar}
\alias{metric_SWAat0to100cm39bar}
\alias{metric_SWAat0to100cm39bar_JJA}
\alias{metric_SWAat0to100cm39bar_daily}
\alias{metric_SWAat20to040cm39bar_JJA}
\alias{metric_SWAat20to040cm39bar_daily}
\alias{metric_SWAat20to100cm39bar_JJA}
\alias{metric_SWAat20to100cm39bar_daily}
\alias{metric_SWAat20to100cm39bar_dailyClim}
\alias{metric_SWAat40to060cm39bar_JJA}
\alias{metric_SWAat40to060cm39bar_daily}
\alias{metric_SWAat60to080cm39bar_JJA}
\alias{metric_SWAat60to080cm39bar_daily}
\alias{metric_SWAat80to100cm39bar_JJA}
\alias{metric_SWAat80to100cm39bar_daily}
\alias{metric_SWP_SoilLayers_MeanMonthly}
\alias{metric_SWPat0to020cm_dailyClim}
\alias{metric_SWPat20to100cm_dailyClim}
\alias{metric_SemiDryDuration_Annual_top50cm}
\alias{metric_SemiDryDuration_Annual_wholeprofile}
\alias{metric_TDDat5C}
\alias{metric_TDDat5C_quarterly}
\alias{metric_TEPET_daily}
\alias{metric_TemperatureMax_Seasonal}
\alias{metric_TemperatureMean_Seasonal}
\alias{metric_TemperatureMin_Seasonal}
\alias{metric_Tmean_JJA}
\alias{metric_Tmean_daily}
\alias{metric_Tmean_dailyClim}
\alias{metric_Tmean_monthly}
\alias{metric_Tmean_monthlyClim}
\alias{metric_TranspirationPeaks_v6}
\alias{metric_TranspirationSeasonality_v1}
\alias{metric_TranspirationSeasonality_v2}
\alias{metric_TranspirationSeasonality_v3}
\alias{metric_TranspirationSeasonality_v4}
\alias{metric_TranspirationSeasonality_v5}
\alias{metric_TranspirationSeasonality_v6}
\alias{metric_Transpiration_Seasonal}
\alias{metric_VWC_Seasonal_top50cm}
\alias{metric_VWC_Seasonal_wholeprofile}
\alias{metric_WDDat5C0to020cm15bar_quarterly}
\alias{metric_WDDat5C0to100cm15bar}
\alias{metric_WDDat5C0to100cm15bar_quarterly}
\alias{metric_WetSoilDays_Seasonal_top50cm}
\alias{metric_WetSoilDays_Seasonal_wholeprofile}
\alias{metric_land_cover_v1}
\alias{metric_land_cover_v2}
\alias{metric_land_cover_v3}
\alias{metric_veg_biomass_annual_v1}
\alias{metric_veg_biomass_annual_v2}
\alias{metric_veg_biomass_annual_v3}
\alias{metric_veg_biomass_monthly_v1}
\alias{metric_veg_biomass_monthly_v2}
\alias{metric_veg_biomass_monthly_v3}
\title{End-user functions that calculate a specific \code{metric}}
\arguments{
\item{path}{A character string. The path to the simulation project folder
that contains the individual folder of each simulated site.}

\item{name_sw2_run}{A character string. The name of the folder
(or zip archive, see \code{zipped_runs}) of the simulated site
for which metrics are to be calculated.
\pkg{rSOILWAT2} input and output is organized following conventions of
\pkg{rSFSW2}, i.e., inputs for each scenario are stored in a list object
named \var{\dQuote{swRunScenariosData}} which is stored
on disk as a file \var{\dQuote{sw_input.RData}};
and output data is stored for each scenario separately in an object
named \var{\dQuote{runDataSC}}
in a file \var{\dQuote{sw_output_scX.RData}} where
\code{X} is the number of the scenario.}

\item{zipped_runs}{A logical value. Describes whether \code{name_sw2_run}
is a zip archive or a regular folder.}

\item{id_scen_used}{An integer vector. The numbers of scenarios for which
metrics are to be calculated.}

\item{list_years_scen_used}{A list of integer vectors.
Each scenario in \code{id_scen_used} must have a corresponding vector of
calendar years (for which the metrics) will be calculated.}

\item{out}{A character string. Signaling whether the functions returns
a time series of yearly values or an aggregate (e.g., mean) across years.
One of \var{\dQuote{ts_years}} or \var{\dQuote{across_years}}.}

\item{soils}{A named list with soil parameters \code{"depth_cm"},
\code{"sand_frac"}, \code{"clay_frac"}, and \code{"gravel_content"}
as numeric vectors with values for each soil layer.}

\item{used_depth_range_cm}{A numeric vector of length two (or \code{NULL}).
The upper and lower depth limit of the zone (depth range) to consider.}

\item{swrcp_and_usage}{A named list with usage of and, if suitable,
parameters of soil water retention curve;
e.g., the output of \code{\link[=load_swrcp_and_usage]{load_swrcp_and_usage()}}.}

\item{...}{Additional arguments}
}
\value{
\itemize{
\item A \code{data.frame} with a least two columns
"site" (identifying sites) and
"group"
(identifying different variables, soil layers, or sub-annual time steps)
\item If \code{do_collect_inputs}, then
a \code{data.frame} where rows represent sites (identified by column "site")
and columns represent soil layers
(columns \code{varname_LX} where X is the soil layer number).
\item Otherwise (i.e., for regular metrics), a \code{data.frame}:
\itemize{
\item rows represent combinations of sites
(identified by column "site") and "groups"
(identified by column "group" for different variables, soil layers,
or sub-annual time steps:
seasonal \var{\dQuote{season}}, quarterly \var{\dQuote{Q}},
monthly \var{\dQuote{mon}}, daily \var{\dQuote{doy}});
\item values of annual time series for different scenarios are present,
if \code{is_out_ts}, in columns with names:
\code{scX_YYYY} (where X is the scenario number and \code{YYYY} is the
calendar year);
\item values of across-year aggregations
are present if the metric itself calculates them or if they were
added by the option \code{add_aggs_across_yrs}, in columns with names
\code{fun_scX_YYYY-ZZZZ} (where "fun" is the aggregation function,
X is the scenario, \code{YYYY} the start year and
\code{ZZZZ} the end year of the period
across which the aggregation was calculated)
}
}
}
\description{
These functions have a name that starts with \code{metric_}.
They have at least the following arguments:
\code{path}, \code{name_sw2_run}, \code{zipped_runs}, \code{id_scen_used},
\code{list_years_scen_used}, \code{out}, and \code{...}.
}
\section{Details}{

Each metric function produces a set of variables (one to many) at a
time step that is specified in the name of the metric function. The default
time step is annual (which may be omitted from the name of the function);
sub-annual time steps include (case-insensitive):
\code{Seasonal} (1 to 12 seasons per year); \code{quarterly}; \code{monthly}; and \code{daily}.


The presence of arguments \code{soils} and \code{swrcp_and_usage} indicates that
the metric function requires as input soil properties or information on the
soil water retention curve respectively.
}

\section{Notes}{

The metric \code{\link[=metric_SW2toTable_daily]{metric_SW2toTable_daily()}} is an exception and produces
a different type of output, i.e., spreadsheets of the most important
daily \code{SOILWAT2} variables, see \code{\link{SW2toTable}}.
}

\section{List of currently available metrics:}{
 \itemize{
   \item \code{\link{metric_AI}}
  \item \code{\link{metric_CWD}}
  \item \code{\link{metric_Climate_annual}}
  \item \code{\link{metric_Climate_monthly}}
  \item \code{\link{metric_Climate_quarterly}}
  \item \code{\link{metric_CorTP_Annual}}
  \item \code{\link{metric_CorTempPPT}}
  \item \code{\link{metric_DDDat5C0to020cm30bar_quarterly}}
  \item \code{\link{metric_DDDat5C0to030cm30bar}}
  \item \code{\link{metric_DDDat5C0to100cm30bar}}
  \item \code{\link{metric_DDDat5C0to100cm30bar_quarterly}}
  \item \code{\link{metric_DR_JJA}}
  \item \code{\link{metric_DR_annual}}
  \item \code{\link{metric_DR_daily}}
  \item \code{\link{metric_DR_monthly}}
  \item \code{\link{metric_DSIat0to100cm15bar}}
  \item \code{\link{metric_DSIat0to100cm30bar}}
  \item \code{\link{metric_DrySoilDays_Seasonal_top50cm}}
  \item \code{\link{metric_DrySoilDays_Seasonal_wholeprofile}}
  \item \code{\link{metric_ET_annual}}
  \item \code{\link{metric_ET_monthly}}
  \item \code{\link{metric_EcologicalDroughtMetrics2023_annual}}
  \item \code{\link{metric_EcologicalDroughtMetrics2023_annualClim}}
  \item \code{\link{metric_Evaporation_Seasonal}}
  \item \code{\link{metric_ExtremeShortTermDryStress_Seasonal_top50cm}}
  \item \code{\link{metric_ExtremeShortTermDryStress_Seasonal_wholeprofile}}
  \item \code{\link{metric_FrostDaysAtNeg5C}}
  \item \code{\link{metric_FrostDays_Seasonal}}
  \item \code{\link{metric_NonDrySWA_Seasonal_top50cm}}
  \item \code{\link{metric_NonDrySWA_Seasonal_wholeprofile}}
  \item \code{\link{metric_PET_Seasonal}}
  \item \code{\link{metric_PPT_JJA}}
  \item \code{\link{metric_PPT_Seasonal}}
  \item \code{\link{metric_PPT_daily}}
  \item \code{\link{metric_PPT_dailyClim}}
  \item \code{\link{metric_PPT_monthlyClim}}
  \item \code{\link{metric_RR2022predictors_annual}}
  \item \code{\link{metric_RR2022predictors_annualClim}}
  \item \code{\link{metric_Radiation_annual}}
  \item \code{\link{metric_Radiation_monthly}}
  \item \code{\link{metric_RecruitmentIndex_v4}}
  \item \code{\link{metric_RecruitmentIndex_v5}}
  \item \code{\link{metric_SMTRs}}
  \item \code{\link{metric_SW2toTable_daily}}
  \item \code{\link{metric_SWA_Seasonal_top50cm}}
  \item \code{\link{metric_SWA_Seasonal_wholeprofile}}
  \item \code{\link{metric_SWAat0to020cm39bar_JJA}}
  \item \code{\link{metric_SWAat0to020cm39bar_daily}}
  \item \code{\link{metric_SWAat0to020cm39bar_dailyClim}}
  \item \code{\link{metric_SWAat0to020to100cm39bar_quarterly}}
  \item \code{\link{metric_SWAat0to100cm30bar}}
  \item \code{\link{metric_SWAat0to100cm39bar}}
  \item \code{\link{metric_SWAat0to100cm39bar_JJA}}
  \item \code{\link{metric_SWAat0to100cm39bar_daily}}
  \item \code{\link{metric_SWAat20to040cm39bar_JJA}}
  \item \code{\link{metric_SWAat20to040cm39bar_daily}}
  \item \code{\link{metric_SWAat20to100cm39bar_JJA}}
  \item \code{\link{metric_SWAat20to100cm39bar_daily}}
  \item \code{\link{metric_SWAat20to100cm39bar_dailyClim}}
  \item \code{\link{metric_SWAat40to060cm39bar_JJA}}
  \item \code{\link{metric_SWAat40to060cm39bar_daily}}
  \item \code{\link{metric_SWAat60to080cm39bar_JJA}}
  \item \code{\link{metric_SWAat60to080cm39bar_daily}}
  \item \code{\link{metric_SWAat80to100cm39bar_JJA}}
  \item \code{\link{metric_SWAat80to100cm39bar_daily}}
  \item \code{\link{metric_SWP_SoilLayers_MeanMonthly}}
  \item \code{\link{metric_SWPat0to020cm_dailyClim}}
  \item \code{\link{metric_SWPat20to100cm_dailyClim}}
  \item \code{\link{metric_SemiDryDuration_Annual_top50cm}}
  \item \code{\link{metric_SemiDryDuration_Annual_wholeprofile}}
  \item \code{\link{metric_TDDat5C}}
  \item \code{\link{metric_TDDat5C_quarterly}}
  \item \code{\link{metric_TEPET_daily}}
  \item \code{\link{metric_TemperatureMax_Seasonal}}
  \item \code{\link{metric_TemperatureMean_Seasonal}}
  \item \code{\link{metric_TemperatureMin_Seasonal}}
  \item \code{\link{metric_Tmean_JJA}}
  \item \code{\link{metric_Tmean_daily}}
  \item \code{\link{metric_Tmean_dailyClim}}
  \item \code{\link{metric_Tmean_monthly}}
  \item \code{\link{metric_Tmean_monthlyClim}}
  \item \code{\link{metric_TranspirationPeaks_v6}}
  \item \code{\link{metric_TranspirationSeasonality_v1}}
  \item \code{\link{metric_TranspirationSeasonality_v2}}
  \item \code{\link{metric_TranspirationSeasonality_v3}}
  \item \code{\link{metric_TranspirationSeasonality_v4}}
  \item \code{\link{metric_TranspirationSeasonality_v5}}
  \item \code{\link{metric_TranspirationSeasonality_v6}}
  \item \code{\link{metric_Transpiration_Seasonal}}
  \item \code{\link{metric_VWC_Seasonal_top50cm}}
  \item \code{\link{metric_VWC_Seasonal_wholeprofile}}
  \item \code{\link{metric_WDDat5C0to020cm15bar_quarterly}}
  \item \code{\link{metric_WDDat5C0to100cm15bar}}
  \item \code{\link{metric_WDDat5C0to100cm15bar_quarterly}}
  \item \code{\link{metric_WetSoilDays_Seasonal_top50cm}}
  \item \code{\link{metric_WetSoilDays_Seasonal_wholeprofile}}
  \item \code{\link{metric_land_cover_v1}}
  \item \code{\link{metric_land_cover_v2}}
  \item \code{\link{metric_land_cover_v3}}
  \item \code{\link{metric_veg_biomass_annual_v1}}
  \item \code{\link{metric_veg_biomass_annual_v2}}
  \item \code{\link{metric_veg_biomass_annual_v3}}
  \item \code{\link{metric_veg_biomass_monthly_v1}}
  \item \code{\link{metric_veg_biomass_monthly_v2}}
  \item \code{\link{metric_veg_biomass_monthly_v3}} 
}}
