% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractions_Main.R
\name{process_values_one_site}
\alias{process_values_one_site}
\title{Extract the requested metrics from one simulation}
\usage{
process_values_one_site(
  fun,
  fun_args,
  name_sw2_run,
  name_sw2_run_soils = NULL,
  is_soils_input = FALSE,
  soils = NULL,
  soil_variables = NULL
)
}
\description{
Extract the requested metrics from one simulation
}
