% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2metrics-defunct.R
\name{rSW2metrics-defunct}
\alias{rSW2metrics-defunct}
\alias{metric_DR}
\alias{metric_ETyr}
\alias{metric_ET}
\alias{metric_Radiation}
\alias{metric_veg_biomass_v1}
\alias{metric_veg_biomass_v2}
\alias{metric_Tmean_monthly_clim}
\alias{metric_TemperatureMean_MeanMonthly}
\alias{metric_PPT_MeanMonthly}
\title{Defunct functions}
\usage{
metric_DR(...)

metric_ETyr(...)

metric_ET(...)

metric_Radiation(...)

metric_veg_biomass_v1(...)

metric_veg_biomass_v2(...)

metric_Tmean_monthly_clim(...)

metric_TemperatureMean_MeanMonthly(...)

metric_PPT_MeanMonthly(...)
}
\arguments{
\item{...}{Replacing old arguments.}
}
\description{
Defunct functions
}
\section{Notes}{

\code{list_all_metrics()} uses the only argument \code{...}
to recognize a metric as defunct.
}

