% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{shorten_run_names}
\alias{shorten_run_names}
\title{Shorten full names of \pkg{rSFSW2} simulation runs}
\usage{
shorten_run_names(run_names, element_sep = "_", N_discard = 2)
}
\arguments{
\item{run_names}{A vector of character strings.}

\item{element_sep}{A character value. The separating character.}

\item{N_discard}{An integer value. The number of elements separated by
\code{element_sep} to discard.}
}
\description{
Full names of \pkg{rSFSW2} simulation runs are composed of three parts
\itemize{
  \item a run number
  \item experimental treatment
  \item site label
}
and concatenated by \var{"_"} as separator, for instance,
\var{"001_Default_SiteA"}, \var{"002_Default_SiteB"}, etc.
}
\details{
It may be desirable to shorten the full names in cases where only none/one
experimental treatment was used, for instance,
\var{SiteA"}, \var{"SiteB"}, etc.
}
\examples{
shorten_run_names(c("001_Default_SiteA", "002_Default_SiteB"))

}
