% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_008_SW2toTable.R
\name{SW2toTable}
\alias{SW2toTable}
\alias{metric_SW2toTable_daily}
\title{Spreadsheet-like daily output of \pkg{SOILWAT2} simulations}
\usage{
metric_SW2toTable_daily(
  path,
  name_sw2_run,
  id_scen_used,
  list_years_scen_used,
  out = "across_years",
  dir_out_SW2toTable = file.path("..", "Outputs"),
  format_share_SW2toTable = c("rds", "csv"),
  outputs_SW2toTable = c("all", "meteo", "snow", "radiation", "waterbalance",
    "evapotranspiration", "soiltemperature", "VWC", "SWP", "SWAat30bar"),
  share_soillayer_ids = NULL,
  ...
)
}
\arguments{
\item{path}{A character string. The path to the simulation project folder
that contains the individual folder of each simulated site.}

\item{name_sw2_run}{A character string. The name of the folder of the
simulated site for which metrics are to be calculated.
\pkg{rSOILWAT2} input and output is organized following conventions of
\pkg{rSFSW2}, i.e., inputs for each scenario are stored in a list object
named \var{\dQuote{swRunScenariosData}} which is stored
on disk as a file \var{\dQuote{sw_input.RData}};
and output data is stored for each scenario separately in an object
named \var{\dQuote{runDataSC}}
in a file \var{\dQuote{sw_output_scX.RData}} where
\code{X} is the number of the scenario.}

\item{id_scen_used}{An integer vector. The numbers of scenarios for which
metrics are to be calculated.}

\item{list_years_scen_used}{A list of integer vectors.
Each scenario in \code{id_scen_used} must have a corresponding vector of
calendar years (for which the metrics) will be calculated.}

\item{out}{A character string. Signaling whether the functions returns
a time series of yearly values or an aggregate (e.g., mean) across years.
One of \var{\dQuote{ts_years}} or \var{\dQuote{across_years}}.}

\item{dir_out_SW2toTable}{A character string. The path at which output
is stored on disk.}

\item{format_share_SW2toTable}{A character string. The function produces
output files in format of either \var{\dQuote{rds}} via \code{\link[base:readRDS]{base::saveRDS()}}
or \var{\dQuote{csv}} via \code{\link[utils:write.table]{utils::write.csv()}}.}

\item{outputs_SW2toTable}{A vector of character strings.
If \var{\dQuote{all}}, then all implemented output variables are included.}

\item{share_soillayer_ids}{A numeric vector. The indices of simulated
soil layers to include in the output files. If \code{NULL}, then
all simulated soil layers are included.}

\item{...}{Additional arguments}
}
\description{
The main purpose of this function is the side-effect of writing
easy-to-share, daily simulation output of relevant variables to disk.
}
\section{Notes}{

This is not a metric as any of the regular \link{metrics}
produced by the package.
}

\section{Details}{

This function produces files as side-effect which
utilizes a different output mechanism than the regular
metrics provided by the package. To make this possible, a few hacks
are required: \itemize{
\item The argument \code{out} is set to \var{\dQuote{across_years}}
despite the fact that this function produces daily output.
\item The function returns dummy output that is processed by the
regular output mechanism of the package.
}
}

