% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{calc_MDD_daily}
\alias{calc_MDD_daily}
\title{Degree-days based on a daily soil moisture, temperature and snow condition}
\usage{
calc_MDD_daily(
  sim_data,
  soils,
  used_depth_range_cm = NULL,
  t_periods = list(op = `>`, limit = 5),
  sm_periods = list(op = `>`, limit = -Inf),
  snow_periods = list(op = `<=`, limit = 0)
)
}
\arguments{
\item{sim_data}{A named list or environment with the following elements
\itemize{
\item \code{"swp_daily"} with two elements
\itemize{
\item \code{"values"}, a list with the named element \code{"swp"},
a two-dimensional object (days by soil layers)
of daily soil water potential \verb{[-bar]}
\item \code{"time"}, a two-dimensional object with days in rows and columns
\code{"Year"} and \code{"Day"} (day of year)
}
\item \code{"temp_daily"}, a list with the element \code{"values"} that contains the
named element \code{"tmean"}, a vector of daily mean air temperatures \verb{[C]}
\item \code{"swe_daily"}, a list with the named element\code{"values"} that contains the
named element \code{"swe"}, a vector of daily water equivalents of
the snow pack \verb{[cm]}
}}

\item{soils}{A named list with soil parameters \code{"depth_cm"},
\code{"sand_frac"}, \code{"clay_frac"}, and \code{"gravel_content"}
as numeric vectors with values for each soil layer.}

\item{used_depth_range_cm}{A numeric vector of length two (or \code{NULL}).
The upper and lower depth limit of the zone (depth range) to consider.}

\item{t_periods}{A list with two named elements that identifies days on
which \code{"tmean"} satisfies the requested conditions,
e.g., \verb{> 5 [C]} for a base temperature of \code{5 [C]}
\itemize{
\item \code{"op"}, a relational operator, e.g., \code{>}
\item \code{"limit"}, a numeric value in units of \verb{[C]}
}}

\item{sm_periods}{A list with two named elements that identifies days on
which \code{"swp"} (converted to \verb{[MPa]}) satisfies the requested conditions.
Note, that dryness (\code{<}, \code{<=}) requires all considered soil layers to be dry,
whereas wetness (\code{>}, \code{>=}) requires at least one considered soil layer
to be moist). For instance, \verb{< -3 [MPa]} identifies dry-degree days
with a dryness below \code{-3 [MPa]}
\itemize{
\item \code{"op"}, a relational operator, e.g., \code{<}
\item \code{"limit"}, a numeric value in units of \verb{[MPa]}
}}

\item{snow_periods}{A list with two named elements that identifies days on
which \code{"swe"} satisfies the requested conditions,
e.g., \verb{<= 0 [mm]} for days without a snow pack
\itemize{
\item \code{"op"}, a relational operator, e.g., \code{<=}
\item \code{"limit"}, a numeric value in units of \verb{[mm]}
}}
}
\value{
A list with two named elements
\itemize{
\item \code{"values"}, a list with the named element \code{"mdd"},
a numeric vector with daily conditioned degree-days \verb{[degC x day]}
\item \code{"time"}, a two-dimensional object with days in rows and columns
\code{"Year"} and \code{"Day"} (day of year), i.e.,
a copy of the input \code{sim_data[["swp_daily"]][["time"]]}
}
}
\description{
Degree-days based on a daily soil moisture, temperature and snow condition
}
\section{Details}{

Argument \code{soils} uses only element \code{"depth_cm"}.
}

\examples{
# Prepare data (here using rSOILWAT2)
swin <- rSOILWAT2::sw_exampleData
soils <- list(depth_cm = rSOILWAT2::swSoils_Layers(swin)[, "depth_cm"])
nSoilLayers <- length(soils[["depth_cm"]])

sim <- rSOILWAT2::sw_exec(swin)
sim_data <- list(
  swp_daily = list(
    time = sim@SWPMATRIC@Day[, c("Year", "Day")],
    values = list(
      swp = sim@SWPMATRIC@Day[, paste0("Lyr_", seq_len(nSoilLayers))]
    )
  ),
  temp_daily = list(
    values = list(tmean = sim@TEMP@Day[, "avg_C"])
  ),
  swe_daily = list(
    values = list(swe = sim@SNOWPACK@Day[, "snowpackWaterEquivalent_cm"])
  )
)

# Dry degree-days "DDDat5C0to100cm30bar" `[C x day]`
# (base 5 C, 0-100 cm, all < -3 MPa, snowfree)
ddd_daily <- calc_MDD_daily(
  sim_data = sim_data,
  soils = soils,
  used_depth_range_cm = c(0, 100),
  t_periods = list(op = `>`, limit = 5),
  sm_periods = list(op = `<`, limit = -3)
)
ddd_annual <- tapply(
  X = ddd_daily[["values"]][["mdd"]],
  INDEX = ddd_daily[["time"]][, "Year"],
  FUN = sum
)

# Wet degree-days "WDDat5C0to100cm15bar" `[C x day]`
# (base 5 C, 0-100 cm, any >-1.5 MPa, snowfree)
wdd_daily <- calc_MDD_daily(
  sim_data = sim_data,
  soils = soils,
  used_depth_range_cm = c(0, 100),
  t_periods = list(op = `>`, limit = 5),
  sm_periods = list(op = `>`, limit = -1.5)
)
wdd_annual <- tapply(
  X = wdd_daily[["values"]][["mdd"]],
  INDEX = wdd_daily[["time"]][, "Year"],
  FUN = sum
)

}
