% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{calc_soillayer_weights}
\alias{calc_soillayer_weights}
\title{Determine widths (as weights) of soil layers within a zone}
\usage{
calc_soillayer_weights(soil_depths_cm, used_depth_range_cm = NULL)
}
\arguments{
\item{soil_depths_cm}{A numeric vector.
The lower depth limits of soil layers in \var{[cm]}.}

\item{used_depth_range_cm}{A numeric vector of length two.
The upper and lower depth limit of the zone (depth range) to consider.}
}
\value{
A numeric vector of the same length as \code{soil_depths_cm}.
  The values correspond to soil layer widths that fall within the
  selected zone. Other values, outside the zone, are set to \code{NA}.
}
\description{
Determine widths (as weights) of soil layers within a zone
}
\examples{
calc_soillayer_weights(c(5, 10, 30, 50), NULL)
calc_soillayer_weights(c(5, 10, 30, 50), c(0, 100))
calc_soillayer_weights(c(5, 10, 30, 50), c(0, 40))
calc_soillayer_weights(c(5, 10, 30, 50), c(0, 40))
calc_soillayer_weights(c(5, 10, 30, 50), c(20, 50))

}
