% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{calc_SWA_mm}
\alias{calc_SWA_mm}
\title{Calculate soil water availability \var{SWA}}
\usage{
calc_SWA_mm(
  sim_swc_daily,
  soils,
  used_depth_range_cm = NULL,
  SWP_limit_MPa = -Inf,
  method = c("across_profile", "by_layer")
)
}
\arguments{
\item{sim_swc_daily}{A named list with a "time" element and a
"values" element containing daily \var{"swc"} for each soil layer
in units of centimeters.}

\item{soils}{A named list with soil parameters \var{"depth_cm"},
\var{"sand_frac"},\var{"clay_frac"}, and \var{"gravel_content"}
as numeric vectors with values for each soil layer.}

\item{used_depth_range_cm}{A numeric vector of length two.}

\item{SWP_limit_MPa}{A numeric value.}

\item{method}{A character string.}
}
\value{
A list with elements "time" and "values" where values represent
available soil water in units of millimeters above \code{SWP_limit_MPa}:
if \code{method} is \var{\dQuote{across_profile}},
then summed across \code{used_depth_range_cm},
if \code{method} is \var{\dQuote{by_layer}},
then columns contain values for each soil layer within
\code{used_depth_range_cm}.
}
\description{
We define available soil water \code{SWA} as the amount of soil water \code{SWC}
that exceeds some base (= critical) amount of soil water \code{SWC_crit[i]},
i.e.,
\deqn{SWA[t,i] = max{0, SWC[t,i] - SWC_crit[i]}}
for day t and soil layer i.
}
\section{Details}{

The base (= critical) amount of soil water \code{SWC_crit[i]} is specified via a
critical soil water potential \code{SWP_crit}, here \code{SWP_limit_MPa}, e.g.,
\verb{SWP_crit = -3.0 MPa}.
The water release curve employed during the simulation run is used to
translate \code{SWP} into volumetric water content \code{VWC}, i.e.,
\deqn{VWC_crit[i,matric] = f(SWP_crit, SWRCp[i])}
where \code{SWRCp} are the parameters describing the water release curve.
However, the translation is only valid for the matric soil, i.e.,
the component without coarse fragments.

\code{SWA} in the presence of coarse fragments is calculated as

\deqn{SWA[t,i] = max{0, SWC[t,i] - w[i] * (1 - cfrag[i]) * VWC_crit[i,matric]}}
or equivalently
\deqn{SWA[t,i] = max{0, w[i] * (1 - cfrag[i]) * (VWC[t,i,matric] - VWC_crit[i,matric])}}

where matric volumetric water \code{VWC} is multiplied by \code{w[i] * (1 - cfrag[i])}
to calculate the amount of water in soil layer i,
correcting for the volume occupied by coarse fragments.

For day t, soil layer i, soil layer width \code{w[i]}, and
\code{cfrag[i]} = fraction of coarse fragments.
}

