% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{determine_used_soillayers}
\alias{determine_used_soillayers}
\title{Determine positions of soil layers within a zone}
\usage{
determine_used_soillayers(
  soil_depths_cm,
  used_depth_range_cm = NULL,
  n_slyrs_has = NULL
)
}
\arguments{
\item{soil_depths_cm}{A numeric vector.
The lower depth limits of soil layers in \var{[cm]}.}

\item{used_depth_range_cm}{A numeric vector of length two.
The upper and lower depth limit of the zone (depth range) to consider.}

\item{n_slyrs_has}{An integer value. The number of simulated soil layers
(optional). The code throws an error if there are fewer soil layers
than selected by \code{used_depth_range_cm} from \code{soil_depths_cm}.}
}
\value{
An integer vector.
}
\description{
Determine positions of soil layers within a zone
}
\examples{
determine_used_soillayers(c(5, 10, 30, 50))
determine_used_soillayers(c(5, 10, 30, 50), c(0, 100))
determine_used_soillayers(c(5, 10, 30, 50), c(0, 40))
determine_used_soillayers(c(5, 10, 30, 50), c(20, 40))
determine_used_soillayers(c(5, 10, 30, 50), c(20, 50))
determine_used_soillayers(c(5, 10, 30, 50), c(20, 30), 3)
\dontrun{
determine_used_soillayers(c(5, 10, 30, 50), c(20, 50), 3)
}

}
