% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{EcologicalDroughtMetrics2023_annual}
\alias{EcologicalDroughtMetrics2023_annual}
\alias{metric_EcologicalDroughtMetrics2023_annual}
\title{Annual time series of ecological drought metrics}
\usage{
metric_EcologicalDroughtMetrics2023_annual(
  path,
  name_sw2_run,
  id_scen_used,
  list_years_scen_used,
  out = c("ts_years", "raw"),
  zipped_runs = FALSE,
  soils,
  ...
)
}
\arguments{
\item{path}{A character string. The path to the simulation project folder
that contains the individual folder of each simulated site.}

\item{name_sw2_run}{A character string. The name of the folder
(or zip archive, see \code{zipped_runs}) of the simulated site
for which metrics are to be calculated.
\pkg{rSOILWAT2} input and output is organized following conventions of
\pkg{rSFSW2}, i.e., inputs for each scenario are stored in a list object
named \var{\dQuote{swRunScenariosData}} which is stored
on disk as a file \var{\dQuote{sw_input.RData}};
and output data is stored for each scenario separately in an object
named \var{\dQuote{runDataSC}}
in a file \var{\dQuote{sw_output_scX.RData}} where
\code{X} is the number of the scenario.}

\item{id_scen_used}{An integer vector. The numbers of scenarios for which
metrics are to be calculated.}

\item{list_years_scen_used}{A list of integer vectors.
Each scenario in \code{id_scen_used} must have a corresponding vector of
calendar years (for which the metrics) will be calculated.}

\item{out}{A character string. Signaling whether the functions returns
a time series of yearly values or an aggregate (e.g., mean) across years.
One of \var{\dQuote{ts_years}} or \var{\dQuote{across_years}}.}

\item{zipped_runs}{A logical value. Describes whether \code{name_sw2_run}
is a zip archive or a regular folder.}

\item{soils}{A named list with soil parameters \code{"depth_cm"},
\code{"sand_frac"}, \code{"clay_frac"}, and \code{"gravel_content"}
as numeric vectors with values for each soil layer.}

\item{...}{Additional arguments}
}
\value{
A return object where \code{group} contains the following
annual variables:
\itemize{
\item Aridity index \verb{[mm / mm]}: \code{"AI"}
(where \code{AI = PPT / PET})
\item Potential evapotranspiration amount \verb{[mm]}: \code{"PET"}
\item Mean daily air temperature \verb{[C]}: \code{"Tmean"}
\item Minimum daily air temperature \verb{[C]}: \code{"Tmin"}
\item Maximum daily air temperature \verb{[C]}: \code{"Tmax"}
\item Precipitation amount \verb{[mm]}: \code{"PPT"}
\item Seasonal timing of precipitation \verb{[-]}: \code{"PPTsst"}
(where \verb{PPTsst = cor(monthly PPT, monthly Tmean)})
\item Total growing degree days \verb{[C x day]}: \code{"TDD"}
\item Seasonal variability of total growing degree days \verb{[C x day / C x day]}: \code{"TDDssv"}
(where \verb{TDDssv = mean(monthly TDD) / sd(monthly TDD)})
\item Warm season length \verb{[day]}: \code{"GrowingSeasonDuration_(mean)|(cv)"}
(where \code{GrowingSeasonDuration} is the longest spell of days with a positive \code{TDD})
\item Evapotranspiration amount \verb{[mm]}: \code{"ET"}
\item Climatic water deficit amount \verb{[mm]}: \code{"CWD"}
(where \code{CWD = PPT - ET})
\item 10-day extreme climatic water deficit \verb{[mm]}: \code{"CWDextreme010d"}
(where \code{CWDextreme10d} is the largest daily \code{CWD} averaged over 10-day periods)
\item Seasonal variability of climatic water deficit \verb{[mm / mm]}: \code{"CWDssv"}
(where \verb{CWDssv = mean(monthly CWD) / sd(monthly CWD)})
\item Seasonal timing of climatic water deficit \verb{[-]}: \code{"CWDsst_"}
(where \verb{CWDsst = cor(monthly CWD, monthly Tmean)})
\item Dry-degree days (0-100 cm, <-3 MPa) \verb{[C x day]}: \code{"DDD"}
\item Dry-degree days of longest spell \verb{[C x day]}: \code{"DDDmaxSpell"}
\item Wet-degree days (0-100 cm, any >-1.5 MPa) \verb{[C x day]}: \code{"WDD"}
\item Seasonal timing of wet-degree days \verb{[-]}: \code{"WDDsst"}
(where \verb{WDDsst = cor(monthly WDD, monthly Tmean)})
\item Mean spell length of dry soils (0-100 cm, <-1.5 MPa) \verb{[day]}: \code{"DSI"}
\item Number of dry soils intervals \verb{[#]}: \code{"nDSI"}
\item First fall frost (<-5 C) \verb{[day of year]}: \code{"FrostFallFirst"}
\item Last spring frost (<-5 C) \verb{[day of year]}: \code{"FrostSpringLast"}
\item Fall recruitment onset \verb{[day of year]}: \code{"RecruitmentFallOnset"}
\item Fall recruitment duration \verb{[day]}: \code{"RecruitmentFallDuration"}
\item Fall recruitment wet-degree days \verb{[C x day]}: \code{"RecruitmentFallWDD"}
\item Spring recruitment onset \verb{[day of year]}: \code{"RecruitmentSpringOnset"}
\item Spring recruitment duration \verb{[day]}: \code{"RecruitmentSpringDuration_"}
\item Spring recruitment wet-degree days \verb{[C x day]}: \code{"RecruitmentSpringWDD"}
\item Available soil moisture (0-100 cm, >-3.9 MPa) \verb{[mm]}: \code{"SWA"}
\item Seasonal variability of available soil moisture \verb{[mm / mm]}: \code{"SWAssv"}
(where \verb{SWAssv = mean(monthly SWA) / sd(monthly CWD)})
\item Seasonal timing of available soil moisture \verb{[-]}: \code{"SWAsst"}
(where \verb{SWAsst = cor(monthly SWA, monthly Tmean)})
}
}
\description{
Annual time series of ecological drought metrics
}
\section{Details}{

The following functions produce all metrics used by Chenoweth et al.:
\itemize{
\item climate metrics
\itemize{
\item \code{metric_Climate_annual()}
\item \code{metric_AI()}
}
\item overall conditions
\itemize{
\item \code{metric_CWD()}
\item \code{metric_SWAat0to100cm39bar()}
\item \code{metric_TDDat5C()}
\item \code{metric_DDDat5C0to100cm30bar()}
\item \code{metric_WDDat5C0to100cm15bar()}
\item \code{metric_FrostDaysAtNeg5C()}
}
\item extreme drought
\itemize{
\item \code{metric_CWD()}
\item \code{metric_DDDat5C0to100cm30bar()}
\item \code{metric_DSIat0to100cm15bar()}
\item \code{metric_DSIat0to100cm15bar()}
}
\item seasonal timing
\itemize{
\item \code{metric_CorTempPPT()}
\item \code{metric_CWD()}
\item \code{metric_SWAat0to100cm39bar()}
\item \code{metric_WDDat5C0to100cm15bar()}
}
\item seasonal variability
\itemize{
\item \code{metric_CWD()}
\item \code{metric_SWAat0to100cm39bar()}
\item \code{metric_TDDat5C()}
}
}
}

\references{
Chenoweth et al. (2023)
Ecologically relevant moisture and temperature metrics for assessing
dryland ecosystem dynamics.
Ecohydrology, 16(3), e2509. \url{https://doi.org/10.1002/eco.2509}
}
\seealso{
\code{\link[=EcologicalDroughtMetrics2023_annualClim]{EcologicalDroughtMetrics2023_annualClim()}};
see \link{metrics} for a complete list of available metric functions.
}
