% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractions_Main.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Main function that organizes the extraction of metrics of a simulation
project and stores them on disk}
\usage{
extract_metrics(args)
}
\description{
Main function that organizes the extraction of metrics of a simulation
project and stores them on disk
}
