% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_AcrossYears.R
\name{fun_across_years}
\alias{fun_across_years}
\alias{mean_cv}
\alias{mean_cv_trend}
\alias{mean_cv_mmk}
\title{Calculate aggregate statistics of values across years}
\usage{
mean_cv(x, na.rm = TRUE)

mean_cv_trend(x, na.rm = TRUE)

mean_cv_mmk(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical value}
}
\description{
Calculate aggregate statistics of values across years
}
