% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Supporting.R
\name{load_swrcp_and_usage}
\alias{load_swrcp_and_usage}
\title{Load soil water retention curve information from a \code{rSOILWAT2} input object}
\usage{
load_swrcp_and_usage(sw_in)
}
\arguments{
\item{sw_in}{A \code{rSOILWAT2} input object of class \code{"swInputData"}.}
}
\value{
A named list including
\itemize{
\item \code{"use_swrc_v6"}, a logical vector that is \code{TRUE}
if the installed version of the \code{rSOILWAT2} package is version \verb{>= 6.0.0}
and if the the simulation input object \code{sw_in} was created by \code{rSOILWAT2}
version \verb{>= 6.0.0} (or if it can be updated to such a version);
\item \code{"swrc_name"}, the name of the soil water retention curve selected in
\code{sw_in} (only used if \code{"use_swrc_v6"}), see \code{\link[rSOILWAT2:swrc_names]{rSOILWAT2::swrc_names()}};
\item \code{"swrcp"}, a matrix with parameters of the selected soil water retention
curve (only used if \code{"use_swrc_v6"}), see \code{\link[rSOILWAT2:SWRCs]{rSOILWAT2::SWRCs}}.
}
}
\description{
Available functionality depends on the currently installed version
of \code{rSOILWAT2} and of the version of \code{rSOILWAT2} that was used
to create the simulation input object \code{sw_in}.
}
\seealso{
\code{\link[=convert_with_swrc]{convert_with_swrc()}}
}
