% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2metrics-package.R
\name{metrics}
\alias{metrics}
\alias{metric_AI}
\alias{metric_Climate_Annual}
\alias{metric_Climate_quarterly}
\alias{metric_CorTempPPT}
\alias{metric_CorTP_Annual}
\alias{metric_CWD}
\alias{metric_DDDat5C0to020cm30bar_quarterly}
\alias{metric_DDDat5C0to030cm30bar}
\alias{metric_DDDat5C0to100cm30bar}
\alias{metric_DDDat5C0to100cm30bar_quarterly}
\alias{metric_DR}
\alias{metric_DR_daily}
\alias{metric_DR_JJA}
\alias{metric_DrySoilDays_Seasonal_top50cm}
\alias{metric_DrySoilDays_Seasonal_wholeprofile}
\alias{metric_DSIat0to100cm15bar}
\alias{metric_DSIat0to100cm30bar}
\alias{metric_ET}
\alias{metric_ETyr}
\alias{metric_Evaporation_Seasonal}
\alias{metric_ExtremeShortTermDryStress_Seasonal_top50cm}
\alias{metric_ExtremeShortTermDryStress_Seasonal_wholeprofile}
\alias{metric_FrostDays_Seasonal}
\alias{metric_FrostDaysAtNeg5C}
\alias{metric_land_cover_v2}
\alias{metric_NonDrySWA_Seasonal_top50cm}
\alias{metric_NonDrySWA_Seasonal_wholeprofile}
\alias{metric_PET_Seasonal}
\alias{metric_PPT_daily}
\alias{metric_PPT_JJA}
\alias{metric_PPT_MeanMonthly}
\alias{metric_PPT_Seasonal}
\alias{metric_Radiation}
\alias{metric_RecruitmentIndex_v4}
\alias{metric_RecruitmentIndex_v5}
\alias{metric_SemiDryDuration_Annual_top50cm}
\alias{metric_SemiDryDuration_Annual_wholeprofile}
\alias{metric_SMTRs}
\alias{metric_SWA_Seasonal_top50cm}
\alias{metric_SWA_Seasonal_wholeprofile}
\alias{metric_SWAat0to020cm39bar_daily}
\alias{metric_SWAat0to020cm39bar_JJA}
\alias{metric_SWAat0to020to100cm39bar_quarterly}
\alias{metric_SWAat0to100cm30bar}
\alias{metric_SWAat0to100cm39bar}
\alias{metric_SWAat0to100cm39bar_daily}
\alias{metric_SWAat0to100cm39bar_JJA}
\alias{metric_SWAat20to040cm39bar_daily}
\alias{metric_SWAat20to040cm39bar_JJA}
\alias{metric_SWAat20to100cm39bar_daily}
\alias{metric_SWAat20to100cm39bar_JJA}
\alias{metric_SWAat40to060cm39bar_daily}
\alias{metric_SWAat40to060cm39bar_JJA}
\alias{metric_SWAat60to080cm39bar_daily}
\alias{metric_SWAat60to080cm39bar_JJA}
\alias{metric_SWAat80to100cm39bar_daily}
\alias{metric_SWAat80to100cm39bar_JJA}
\alias{metric_SWP_SoilLayers_MeanMonthly}
\alias{metric_TDDat5C}
\alias{metric_TDDat5C_quarterly}
\alias{metric_TemperatureMax_Seasonal}
\alias{metric_TemperatureMean_MeanMonthly}
\alias{metric_TemperatureMean_Seasonal}
\alias{metric_TemperatureMin_Seasonal}
\alias{metric_Tmean_daily}
\alias{metric_Tmean_JJA}
\alias{metric_Transpiration_Seasonal}
\alias{metric_TranspirationPeaks_v6}
\alias{metric_TranspirationSeasonality_v1}
\alias{metric_TranspirationSeasonality_v2}
\alias{metric_TranspirationSeasonality_v3}
\alias{metric_TranspirationSeasonality_v4}
\alias{metric_TranspirationSeasonality_v5}
\alias{metric_TranspirationSeasonality_v6}
\alias{metric_veg_biomass_v1}
\alias{metric_veg_biomass_v2}
\alias{metric_VWC_Seasonal_top50cm}
\alias{metric_VWC_Seasonal_wholeprofile}
\alias{metric_WDDat5C0to020cm15bar_quarterly}
\alias{metric_WDDat5C0to100cm15bar}
\alias{metric_WDDat5C0to100cm15bar_quarterly}
\alias{metric_WetSoilDays_Seasonal_top50cm}
\alias{metric_WetSoilDays_Seasonal_wholeprofile}
\title{End-user functions that return a specific \code{metric}}
\arguments{
\item{path}{A character string. The path to the simulation project folder
that contains the individual folder of each simulated site.}

\item{name_sw2_run}{A character string. The name of the folder of the
simulated site for which metrics are to be calculated.
\pkg{rSOILWAT2} input and output is organized following conventions of
\pkg{rSFSW2}, i.e., inputs for each scenario are stored in a list object
named \var{\dQuote{swRunScenariosData}} which is stored
on disk as a file \var{\dQuote{sw_input.RData}};
and output data is stored for each scenario separately in an object
named \var{\dQuote{runDataSC}}
in a file \var{\dQuote{sw_output_scX.RData}} where
\code{X} is the number of the scenario.}

\item{id_scen_used}{An integer vector. The numbers of scenarios for which
metrics are to be calculated.}

\item{list_years_scen_used}{A list of integer vectors.
Each scenario in \code{id_scen_used} must have a corresponding vector of
calendar years (for which the metrics) will be calculated.}

\item{out}{A character string. Signaling whether the functions returns
a time series of yearly values or an aggregate (e.g., mean) across years.
One of \var{\dQuote{ts_years}} or \var{\dQuote{across_years}}.}

\item{soils}{A named list of numeric vectors. The presence of the
argument \code{soils} indicates that the function in question requires
soil information as inputs.
The named elements include \var{\dQuote{depth_cm}},
\var{\dQuote{sand_frac}}, and \var{\dQuote{clay_frac}} and
contain the respective values for each soil layer at the site.}

\item{...}{Additional arguments}
}
\description{
These functions have a name that starts with \code{metric_}.
They have at least the following arguments:
\code{path}, \code{name_sw2_run}, \code{id_scen_used},
\code{list_years_scen_used}, \code{out}, and \code{...}.
}
\section{List of currently available metrics:}{
 \itemize{
   \item \var{metric_AI}
  \item \var{metric_Climate_Annual}
  \item \var{metric_Climate_quarterly}
  \item \var{metric_CorTempPPT}
  \item \var{metric_CorTP_Annual}
  \item \var{metric_CWD}
  \item \var{metric_DDDat5C0to020cm30bar_quarterly}
  \item \var{metric_DDDat5C0to030cm30bar}
  \item \var{metric_DDDat5C0to100cm30bar}
  \item \var{metric_DDDat5C0to100cm30bar_quarterly}
  \item \var{metric_DR}
  \item \var{metric_DR_daily}
  \item \var{metric_DR_JJA}
  \item \var{metric_DrySoilDays_Seasonal_top50cm}
  \item \var{metric_DrySoilDays_Seasonal_wholeprofile}
  \item \var{metric_DSIat0to100cm15bar}
  \item \var{metric_DSIat0to100cm30bar}
  \item \var{metric_ET}
  \item \var{metric_ETyr}
  \item \var{metric_Evaporation_Seasonal}
  \item \var{metric_ExtremeShortTermDryStress_Seasonal_top50cm}
  \item \var{metric_ExtremeShortTermDryStress_Seasonal_wholeprofile}
  \item \var{metric_FrostDays_Seasonal}
  \item \var{metric_FrostDaysAtNeg5C}
  \item \var{metric_land_cover_v2}
  \item \var{metric_NonDrySWA_Seasonal_top50cm}
  \item \var{metric_NonDrySWA_Seasonal_wholeprofile}
  \item \var{metric_PET_Seasonal}
  \item \var{metric_PPT_daily}
  \item \var{metric_PPT_JJA}
  \item \var{metric_PPT_MeanMonthly}
  \item \var{metric_PPT_Seasonal}
  \item \var{metric_Radiation}
  \item \var{metric_RecruitmentIndex_v4}
  \item \var{metric_RecruitmentIndex_v5}
  \item \var{metric_SemiDryDuration_Annual_top50cm}
  \item \var{metric_SemiDryDuration_Annual_wholeprofile}
  \item \var{metric_SMTRs}
  \item \var{metric_SWA_Seasonal_top50cm}
  \item \var{metric_SWA_Seasonal_wholeprofile}
  \item \var{metric_SWAat0to020cm39bar_daily}
  \item \var{metric_SWAat0to020cm39bar_JJA}
  \item \var{metric_SWAat0to020to100cm39bar_quarterly}
  \item \var{metric_SWAat0to100cm30bar}
  \item \var{metric_SWAat0to100cm39bar}
  \item \var{metric_SWAat0to100cm39bar_daily}
  \item \var{metric_SWAat0to100cm39bar_JJA}
  \item \var{metric_SWAat20to040cm39bar_daily}
  \item \var{metric_SWAat20to040cm39bar_JJA}
  \item \var{metric_SWAat20to100cm39bar_daily}
  \item \var{metric_SWAat20to100cm39bar_JJA}
  \item \var{metric_SWAat40to060cm39bar_daily}
  \item \var{metric_SWAat40to060cm39bar_JJA}
  \item \var{metric_SWAat60to080cm39bar_daily}
  \item \var{metric_SWAat60to080cm39bar_JJA}
  \item \var{metric_SWAat80to100cm39bar_daily}
  \item \var{metric_SWAat80to100cm39bar_JJA}
  \item \var{metric_SWP_SoilLayers_MeanMonthly}
  \item \var{metric_TDDat5C}
  \item \var{metric_TDDat5C_quarterly}
  \item \var{metric_TemperatureMax_Seasonal}
  \item \var{metric_TemperatureMean_MeanMonthly}
  \item \var{metric_TemperatureMean_Seasonal}
  \item \var{metric_TemperatureMin_Seasonal}
  \item \var{metric_Tmean_daily}
  \item \var{metric_Tmean_JJA}
  \item \var{metric_Transpiration_Seasonal}
  \item \var{metric_TranspirationPeaks_v6}
  \item \var{metric_TranspirationSeasonality_v1}
  \item \var{metric_TranspirationSeasonality_v2}
  \item \var{metric_TranspirationSeasonality_v3}
  \item \var{metric_TranspirationSeasonality_v4}
  \item \var{metric_TranspirationSeasonality_v5}
  \item \var{metric_TranspirationSeasonality_v6}
  \item \var{metric_veg_biomass_v1}
  \item \var{metric_veg_biomass_v2}
  \item \var{metric_VWC_Seasonal_top50cm}
  \item \var{metric_VWC_Seasonal_wholeprofile}
  \item \var{metric_WDDat5C0to020cm15bar_quarterly}
  \item \var{metric_WDDat5C0to100cm15bar}
  \item \var{metric_WDDat5C0to100cm15bar_quarterly}
  \item \var{metric_WetSoilDays_Seasonal_top50cm}
  \item \var{metric_WetSoilDays_Seasonal_wholeprofile} 
}}
