% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractions_Main.R
\name{process_arguments}
\alias{process_arguments}
\title{Process arguments of calling script}
\usage{
process_arguments(x)
}
\arguments{
\item{x}{A vector of character strings.
The return value of \code{\link[base]{commandArgs}}.}
}
\value{
A named list with the processed arguments.
}
\description{
Process arguments of calling script
}
