% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{RR2022predictors_annual}
\alias{RR2022predictors_annual}
\alias{metric_RR2022predictors_annual}
\title{Annual time series underlying the 19 R&R predictors}
\usage{
metric_RR2022predictors_annual(
  path,
  name_sw2_run,
  id_scen_used,
  list_years_scen_used,
  out = c("ts_years", "raw"),
  zipped_runs = FALSE,
  soils,
  ...
)
}
\description{
The 19 predictors of resilience & resistance indicators
(see \code{metric_RR2022predictors_annualClim()}) are long-term
means, standard deviations, or coefficients of variation of
annual values.
}
\references{
Chambers et al. (2023)
New indicators of ecological resilience and invasion resistance to support
prioritization and management in the sagebrush biome, United States.
Frontiers in Ecology and Evolution, 10, 1–17.
\url{https://doi.org/10.3389/fevo.2022.1009268}
}
\seealso{
\code{\link[=RR2022predictors_annualClim]{RR2022predictors_annualClim()}} and
\code{\link[=metric_RR2022predictors_annual]{metric_RR2022predictors_annual()}}
}
