% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Metrics_002_newRR.R
\name{RR2022predictors_annualClim}
\alias{RR2022predictors_annualClim}
\alias{metric_RR2022predictors_annualClim}
\title{19 predictors of resilience & resistance indicators}
\usage{
metric_RR2022predictors_annualClim(
  path,
  name_sw2_run,
  id_scen_used,
  list_years_scen_used,
  out = "across_years",
  zipped_runs = FALSE,
  soils,
  ...
)
}
\value{
A return object where \code{group} contains the following annual
variables:
\itemize{
\item \code{"Tmean_mean"}
\item \code{"Trange_diurnal_mean"}
\item \code{"Tmean_coldestmonth_mean"}
\item \code{"Tmean_coldestmonth_sd"}
\item \code{"Tmean_hottestmonth_sd"}
\item \code{"PPT_mean"}
\item \code{"PPT_cv"}
\item \code{"Rain_mean"}
\item \code{"PPTinJAS_mean"}
\item \code{"PPT_driestmonth_mean"}
\item \code{"PET_cv"}
\item \code{"ET_cv"}
\item \code{"CWD_mean"}
\item \code{"CWD_mon_corr_temp_mean"}
\item \code{"CWD_mon_cv_mean"}
\item \code{"DDD_mean"}
\item \code{"DSI_duration_mean"}
\item \code{"CorTempPPT_mean"}
\item \code{"DeepDrainage_mean"}
}
}
\description{
The 19 predictors of resilience & resistance indicators are long-term
means, standard deviations, or coefficients of variation of
annual values (see \code{metric_RR2022predictors_annual()}).
}
\section{Notes}{

\itemize{
\item Argument \code{fun_aggs_across_yrs} is ignored.
\item Values are \code{NA} if any year is requested but was not simulated.
}
}

\references{
Chambers et al. (2023)
New indicators of ecological resilience and invasion resistance to support
prioritization and management in the sagebrush biome, United States.
Frontiers in Ecology and Evolution, 10, 1–17.
\url{https://doi.org/10.3389/fevo.2022.1009268}
}
\seealso{
\code{\link[=RR2022predictors_annual]{RR2022predictors_annual()}} and
\code{\link[=metric_RR2022predictors_annualClim]{metric_RR2022predictors_annualClim()}}
}
