% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2metrics-package.R
\docType{package}
\name{rSW2metrics}
\alias{rSW2metrics}
\alias{rSW2metrics-package}
\title{Package \pkg{rSW2metrics}: Collection of functions to calculate
ecohydrological metrics from output created by
\pkg{rSOILWAT2} or \pkg{rSFSW2}simulation experiments.}
\description{
Collection of functions to calculate ecohydrological metrics
    from output created by rSOILWAT2 or rSFSW2 simulation experiments.
}
\section{Details}{

Recommended setup:
\enumerate{
\item Copy file \var{\dQuote{Project_Parameters.R}}.
Specify values for your specific project.
\item Copy file \var{\dQuote{Script_to_Extract_Metric.R}}.
In most cases, this script can be used without changes.
\item Copy file \var{\dQuote{Script_Shell_Extracting_rSW2metrics.sh}}.
Specify arguments/options and remove calls to metrics not needed.
\item Run the extraction by executing
\var{\dQuote{Script_Shell_Extracting_rSW2metrics.sh}}
on the command line.
}

Example code for copying the three files to your project folder:\preformatted{file.copy(
  from = list.files(
    path = system.file("exec", package = "rSW2metrics"),
    full.names = TRUE
  ),
  to = PATH_TO_YOUR_PROJECT
)
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrylandEcology/rSW2metrics}
  \item Report bugs at \url{https://github.com/DrylandEcology/rSW2metrics/issues}
}

}
\author{
\strong{Maintainer}: Daniel Schlaepfer \email{danielrschlaepfer@gmail.com} (\href{https://orcid.org/0000-0001-9973-2065}{ORCID})

}
